+++ hammer-openfront-extension/entrypoints/hooks.content.ts
@@ -1288,7 +1288,7 @@ export default defineContentScript({
         timeoutIds.length = 0;
         intervalIds.length = 0;
       },
-      version: "15.0.0-ext",
+      version: "15.0.1-ext",
     };
 
     console.log("[Hammer:Main] Hooks installed at document_start");